requirements.txt
langchain
google-generativeai
python-dotenv
PyPDF2
pytesseract
opencv-python
langchain-community
langchain-google-genai
